# docs/swagger.yaml
basePath: /
definitions:
  deck.Card:
    properties:
      suit:
        type: string
      value:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A small HTTP API for creating, persisting, and manipulating decks of
    cards.
  title: go-cards API
  version: "1.0"
paths:
  /card/random:
    get:
      description: Returns a random card from a fresh, unsaved deck.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deck.Card'
      summary: Get a random card
      tags:
      - cards
  /deck:
    post:
      description: Creates a new 52-card deck, persists it to disk, and returns its
        id.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              id:
                type: string
            type: object
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create a new deck
      tags:
      - decks
  /deck/{id}:
    delete:
      description: Removes the stored deck file for the given id.
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: no content
        "404":
          description: deck not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete a stored deck
      tags:
      - decks
    get:
      description: Returns the full stored deck for the given id.
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/deck.Card'
            type: array
        "404":
          description: deck not found
          schema:
            type: string
      summary: Get a stored deck
      tags:
      - decks
  /deck/{id}/card/random:
    get:
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deck.Card'
        "404":
          description: deck not found
          schema:
            type: string
      summary: Get a random card from a stored deck
      tags:
      - decks
  /deck/{id}/card/top:
    get:
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deck.Card'
        "400":
          description: deck is empty
          schema:
            type: string
        "404":
          description: deck not found
          schema:
            type: string
      summary: Get the top card of a stored deck
      tags:
      - decks
  /deck/{id}/deal:
    post:
      description: Deals handSize cards off the top of the stored deck. The stored
        deck shrinks by handSize.
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      - description: Number of cards to deal
        in: query
        name: handSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/deck.Card'
            type: array
        "400":
          description: invalid handSize or handSize exceeds deck size
          schema:
            type: string
        "404":
          description: deck not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Deal cards from a stored deck
      tags:
      - decks
  /deck/{id}/shuffle:
    post:
      description: Shuffles the stored deck in place and returns the shuffled deck.
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/deck.Card'
            type: array
        "404":
          description: deck not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Shuffle a stored deck
      tags:
      - decks
swagger: "2.0"